Removed the [invisible-element] property
For all "create widget event handlers", since if invisible, the widget
should be rendered with the same element as if visible. Only with a
style of "display:none;" added.

Also created some documentation for p5.web (not finished yet though)

And added some more unit tests for [add] for nodes having values that
are nodes themselves.