Fixed severe error in form elements that are invisible
When an element was invisible, either directly, or through its parent
being invisible, and it was a form element (input, textarea, select,
etc) - Then its value would not be persisted, since this was exclusively
fetched from "LoadFormData". Now I check if the element is invisible,
and if it is, I presist its value in ViewState.