Fixed sever style bug
Style cannot use _getChange logic, since it is impossible to make sure
we find the right offset. Hence, when style changes on widget, we
entirely re-render the style property. Which shouldn't be too much of a
big deal, since you're not really supposed to use styles in the first
place, right ...? ;)