Fixed [oninit], and writte more dox
[oninit] does not fire before widget is physically displayed, either
through having its [visible] property set to true, or one of its
ancestor widget's [visible] properties set to true. This has the effect
of not evaluating [oninit] before widget physically becomes visible, and
in fact, *every* time it becomes visible. This is important for the
widget to have support for being inside widgets that are not visible,
and postponing its [oninit] to it becomes visible. This is important, in
for instance, TabControls and such, having complex widgets, including
JavaScript during [oninit].