Lots of changes
* App loader can let app override title of page (make sure you
re-create the app-load CMS page in your database if you need to use
this feature)
* Supports download of files in the CMS with different physical
filenames than that which is supposed to be displayed to the user.
* Minor bugfix in [sys42.windows.modal] that passed in arguments in the
wrong order.
* [.onfinished] lambda callback for [p5.pop3.get], which allows the
user to handle the emails received, before QUIT signal is sent, and if
an exception occurs during handling, no QUIT signal is sent to POP3
server.
* Doesn’t throw exception if no standard MailboxAddress is given in
[p5.pop3.get]
* Unrolls file paths in the MIME creator, allowing for using
“environment variables” for file names/paths.
* p5.mime doesn’t throw if MIME part is neither Multipart nor
MessagePart
* Support for searching for private and public GnuPG keys.
* Long type in p5.types, which is often used in databases and such for
IDs.
* [trim] now doesn’t throw if source is empty.
* DBNull support in MySQL adapter.