Changed the logic for desktop icons
Such that now it only checks for existence of “desktop.hl” to figure
out if it should create an icon or not. This allows us to have URL
resolvers for modules not having an icon, such as for instance the
[micro.download.file] supporting Active Event, that allows the user to
download files, but does not imply that Micro should have its own
desktop icon.