[eval-whitelist] fix
Such that it removes the whitelist when an event is invoked. This
allows you to whitelist an event, which is internally using illegal
events, avoiding having to expose the event that is used internally by
your own events. But it also makes it logically impossible to (safely)
invoke an event that evaluates a lambda object, specified by caller!